Move the rest of BaseRecord to .cpp
